python_ai_service/tools/image_analysis.py
"""
Image Analysis Tool using CLIP and Computer Vision
"""

import os
import base64
from typing import Dict, Any, List
from PIL import Image
import io

class ImageAnalysisTool:
    def __init__(self):
        self.model_loaded = False
        
    async def analyze_image(self, image_data: str, analysis_type: str = "general") -> Dict[str, Any]:
        """
        Analyze image using CLIP or other computer vision models
        
        Args:
            image_data: Base64 encoded image or file path
            analysis_type: Type of analysis (general, objects, text, emotions)
        """
        
        try:
            # Process image
            image = self._process_image(image_data)
            
            if not image:
                return {"error": "Could not process image"}
            
            # Perform analysis based on type
            if analysis_type == "general":
                return await self._general_analysis(image)
            elif analysis_type == "objects":
                return await self._object_detection(image)
            elif analysis_type == "text":
                return await self._text_extraction(image)
            elif analysis_type == "emotions":
                return await self._emotion_analysis(image)
            else:
                return await self._general_analysis(image)
                
        except Exception as e:
            return {"error": f"Image analysis failed: {str(e)}"}
    
    def _process_image(self, image_data: str) -> Image.Image:
        """Process image from base64 or file path"""
        try:
            if image_data.startswith('data:image'):
                # Extract base64 part
                header, encoded = image_data.split(',', 1)
                image_bytes = base64.b64decode(encoded)
                return Image.open(io.BytesIO(image_bytes))
            elif os.path.exists(image_data):
                # File path
                return Image.open(image_data)
            else:
                # Assume it's base64 without header
                image_bytes = base64.b64decode(image_data)
                return Image.open(io.BytesIO(image_bytes))
        except Exception as e:
            print(f"Image processing error: {e}")
            return None
    
    async def _general_analysis(self, image: Image.Image) -> Dict[str, Any]:
        """General image analysis"""
        
        # Mock analysis until CLIP is available
        width, height = image.size
        format_info = image.format or "Unknown"
        mode = image.mode
        
        return {
            "analysis_type": "general",
            "description": "This appears to be a digital image with various visual elements. The composition includes multiple objects and scenes that could be analyzed for content, style, and context.",
            "technical_details": {
                "width": width,
                "height": height,
                "format": format_info,
                "mode": mode,
                "size_kb": len(image.tobytes()) // 1024
            },
            "confidence": 0.85,
            "detected_elements": [
                "Visual content",
                "Digital composition",
                "Color elements",
                "Structured layout"
            ],
            "suggestions": [
                "Image quality is good for analysis",
                "Multiple analysis types available",
                "CLIP integration pending for detailed semantic analysis"
            ]
        }
    
    async def _object_detection(self, image: Image.Image) -> Dict[str, Any]:
        """Object detection in image"""
        
        return {
            "analysis_type": "object_detection",
            "objects": [
                {"name": "object_1", "confidence": 0.9, "bbox": [10, 10, 100, 100]},
                {"name": "object_2", "confidence": 0.8, "bbox": [150, 20, 200, 80]}
            ],
            "total_objects": 2,
            "note": "CLIP object detection integration pending"
        }
    
    async def _text_extraction(self, image: Image.Image) -> Dict[str, Any]:
        """Extract text from image (OCR)"""
        
        return {
            "analysis_type": "text_extraction",
            "extracted_text": "Sample text extracted from image",
            "text_regions": [
                {"text": "Sample text", "confidence": 0.95, "bbox": [50, 50, 200, 70]}
            ],
            "note": "OCR integration pending"
        }
    
    async def _emotion_analysis(self, image: Image.Image) -> Dict[str, Any]:
        """Analyze emotions in image (faces, scenes)"""
        
        return {
            "analysis_type": "emotion_analysis",
            "emotions": [
                {"emotion": "neutral", "confidence": 0.7},
                {"emotion": "positive", "confidence": 0.6}
            ],
            "overall_sentiment": "neutral",
            "note": "Emotion analysis integration pending"
        }

# Global instance
image_analyzer = ImageAnalysisTool()