python_ai_service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.26
langchain-community==0.3.27
chromadb==1.0.15
python-dotenv==1.1.1
requests==2.32.4
pillow==11.3.0
pydantic==2.11.7